python_conversion/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
PyJWT==2.10.1
psycopg2-binary==2.9.10
PyYAML==6.0.2
pytest==8.3.5
flask-login==0.6.3
